requirements.txt
timm==0.9.16
librosa==0.10.1
gensim==4.3.2
fast_pytorch_kmeans